# restapi/internal/api/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Personal Computer Maintenance Management System REST API.",
        "title": "PersonaCMMS API",
        "contact": {
            "email": "greenrivercodelabs@gmail.com"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0",
            "url": "https://spdx.org/licenses/AGPL-3.0-or-later.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/assets/{assetId}/tasks/{taskId}/date-triggers": {
            "get": {
                "description": "List date triggers for a task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "date-triggers"
                ],
                "summary": "List date triggers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.DateTriggerResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a date trigger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "date-triggers"
                ],
                "summary": "Create a date trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Date Trigger object",
                        "name": "dateTrigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DateTriggerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.DateTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/assets/{assetId}/tasks/{taskId}/date-triggers/{dateTriggerId}": {
            "get": {
                "description": "Get a date trigger",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "date-triggers"
                ],
                "summary": "Get a date trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date Trigger Id",
                        "name": "dateTriggerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DateTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update a date trigger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "date-triggers"
                ],
                "summary": "Update a date trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date Trigger Id",
                        "name": "dateTriggerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Date Trigger object",
                        "name": "dateTrigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DateTriggerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DateTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a date trigger",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "date-triggers"
                ],
                "summary": "Delete a date trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date Trigger Id",
                        "name": "dateTriggerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.DateTriggerRequest": {
            "type": "object",
            "properties": {
                "scheduled_date": {
                    "type": "string"
                }
            }
        },
        "api.DateTriggerResponse": {
            "type": "object",
            "properties": {
                "scheduled_date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}